minor bug fix,  1 package renamed
