+++ ShopAdvisor/src/productHandler/Main.java
@@ -1,5 +1,5 @@
 
-package Seller_editor;
+package productHandler;
 
 import javafx.application.Application;
 import javafx.collections.FXCollections;
