+++ ShopAdvisor/src/ui/SellerUI.fxml
@@ -2,22 +2,27 @@
 
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
 <?import javafx.scene.control.ScrollBar?>
 <?import javafx.scene.control.Tab?>
 <?import javafx.scene.control.TabPane?>
 <?import javafx.scene.control.TableColumn?>
 <?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.ToggleButton?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.ColumnConstraints?>
+<?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<BorderPane fx:id="rootPane" prefHeight="637.0" prefWidth="978.0" styleClass="Salesroot" stylesheets="@style.css" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ui.CustomerUIController">
+<BorderPane fx:id="Salesroot" prefHeight="637.0" prefWidth="978.0" styleClass="Salesroot" stylesheets="@style.css" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ui.SellerUIController">
    <center>
       <VBox prefHeight="200.0" prefWidth="100.0" BorderPane.alignment="CENTER">
          <children>
@@ -61,7 +66,66 @@
                 </Tab>
                   <Tab text="Product List">
                     <content>
-                      <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0" />
+                      <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
+                           <children>
+                              <GridPane layoutX="46.0" layoutY="26.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="100.0" AnchorPane.rightAnchor="100.0" AnchorPane.topAnchor="0.0">
+                                <columnConstraints>
+                                  <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+                                  <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+                                </columnConstraints>
+                                <rowConstraints>
+                                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                                </rowConstraints>
+                                 <children>
+                                    <Label alignment="CENTER" contentDisplay="CENTER" prefHeight="63.0" prefWidth="383.0" text="Product Name">
+                                       <font>
+                                          <Font size="44.0" />
+                                       </font>
+                                    </Label>
+                                    <Label alignment="CENTER" contentDisplay="CENTER" layoutX="10.0" layoutY="8.0" prefHeight="63.0" prefWidth="383.0" text="Availability" GridPane.columnIndex="1">
+                                       <font>
+                                          <Font size="44.0" />
+                                       </font>
+                                    </Label>
+                                    <ToggleButton mnemonicParsing="false" prefHeight="63.0" prefWidth="387.0" text="Yes" GridPane.columnIndex="1" GridPane.rowIndex="1">
+                                       <font>
+                                          <Font size="29.0" />
+                                       </font>
+                                    </ToggleButton>
+                                    <ToggleButton layoutX="394.0" layoutY="70.0" mnemonicParsing="false" prefHeight="63.0" prefWidth="387.0" text="Yes" GridPane.columnIndex="1" GridPane.rowIndex="6">
+                                       <font>
+                                          <Font size="29.0" />
+                                       </font>
+                                    </ToggleButton>
+                                    <ToggleButton layoutX="394.0" layoutY="70.0" mnemonicParsing="false" prefHeight="63.0" prefWidth="387.0" text="Yes" GridPane.columnIndex="1" GridPane.rowIndex="5">
+                                       <font>
+                                          <Font size="29.0" />
+                                       </font>
+                                    </ToggleButton>
+                                    <ToggleButton layoutX="394.0" layoutY="70.0" mnemonicParsing="false" prefHeight="63.0" prefWidth="387.0" text="Yes" GridPane.columnIndex="1" GridPane.rowIndex="4">
+                                       <font>
+                                          <Font size="29.0" />
+                                       </font>
+                                    </ToggleButton>
+                                    <ToggleButton layoutX="394.0" layoutY="70.0" mnemonicParsing="false" prefHeight="63.0" prefWidth="387.0" text="Yes" GridPane.columnIndex="1" GridPane.rowIndex="3">
+                                       <font>
+                                          <Font size="29.0" />
+                                       </font>
+                                    </ToggleButton>
+                                    <ToggleButton layoutX="394.0" layoutY="70.0" mnemonicParsing="false" prefHeight="63.0" prefWidth="387.0" text="Yes" GridPane.columnIndex="1" GridPane.rowIndex="2">
+                                       <font>
+                                          <Font size="29.0" />
+                                       </font>
+                                    </ToggleButton>
+                                 </children>
+                              </GridPane>
+                           </children></AnchorPane>
                     </content>
                   </Tab>
                   <Tab text="Change/Edit info">
